requirements.txt
anthropic
rich
tavily-python
ollama
groq
python-dotenv
